Create .gitignore, new view layouts and related components
A new .gitignore file is added to the workspace. Created a new view layout file 'app.blade.php' and its related component 'AppLayouts.php'. Also, updated the 'welcome.blade.php' file with the new layout. This was necessary to structure the views and components of the application in a more organized and scalable manner.